Create component for forgot password confirmation
