+++ src/services/authentication.service.ts
@@ -1,5 +1,6 @@
 import { SupabaseClient } from "@supabase/supabase-js"
 import { Database } from "@/types/database"
+import { toast } from "react-toastify"
 
 export const setName = async (
   supabaseClient: SupabaseClient<Database>,
@@ -11,4 +12,16 @@ export const setName = async (
 
 export const isInMemberDatabase = (supabaseClient: SupabaseClient<Database>, id: string) => {
   return supabaseClient.from("member").select("*", { count: "exact", head: true }).eq("id", id)
+}
+
+export const resetPasswordForEmail = async (supabase: SupabaseClient, email: string) => {
+  const { error } = await supabase.auth.resetPasswordForEmail(email, {
+    redirectTo: `${window.location.origin}/forgot-password/reset`,
+  })
+  if (error) {
+    toast.error(error.message)
+    return false
+  } else {
+    return true
+  }
 }
